backend/requirements.txt
Flask
Flask-Cors
torch
torchvision
Pillow
requests
numpy
opencv-python